requirements.txt
#python 3.9
jupyter
pandas
tqdm
opencv-python
Pillow
joblib
matplotlib
transformers
torch
torchvision
scikit-learn
graphviz
seaborn
shapely
python-dotenv
pydantic-settings
scikit-image
numpy==1.23.5
mxnet==1.9.1
#mxnet-cu112==1.9.1
gluoncv==0.10.5
xgboost
lightgbm
catboost
shap
lime
